fix grid within accordion
